close-api/requirements.txt
customtkinter==5.2.2
requests==2.32.5
pillow==12.1.1